Configurable header
Insert the logic for https://github.com/georchestra/superset-core/issues/46
In geOrchestra, the header (top menu) is controlled by params rom the
default.properties file. This commit provides a means to parse this file
and get a header that is consistent with geOrchestra's other apps.
It is also possible to disable the header's display by setting
GEORCHESTRA_NOHEADER=True in the superset config file