feat: disable EventLogger
By default, Superset uses a DBEventLogger, storing all actions on
Superset, in the DB. Problem is that it can bloat the DB after some time.
And since we now have to allow writing to the event logs to Public
to avoid having error messages (see #fa5df900), writing logs in the DB
doesn't feel that safe anymore.
We can alternatively use the StdOutEventLogger that logs those events
on stdout.
By default, the NullEventLogger doesn't log them at all,
it will be up to the platform admins to see if they want to log them and how.
On the longer term, we'll probably be considering adding back some logging
mechanism for analytics purposes.