fix: CSRF issue https://github.com/georchestra/superset/issues/14
A very short-lived session was configured to hack around the roles update
(if a connected user has their roles changed, it is not detected automatically.
Having a short-lived session ensured it would be updated soon enough). This is
wrong, because it means that the session will be expired while the user is
still active, causing such issues.
This fix handles specifically the roles expiration possibility by checking
on a regular basis. Check frequency can be configured by variable
GEORCHESTRA_ROLES_CHECK_FREQUENCY (in minutes) Default is 5 minutes.